refactoring script to parts
