+++ .gitignore
@@ -4,3 +4,5 @@ venv
 *.local
 .idea
 outputs/*
+*/__pycache__
+*.pyc
