+++ scripts/views/PygameView/SdPaint.py
@@ -1,0 +1,1270 @@
+import functools
+import gc
+import os
+
+import pygame
+import pygame.gfxdraw
+import requests
+import threading
+# import numpy as np
+import base64
+import io
+import json
+import time
+import math
+from PIL import Image, ImageOps
+import tkinter as tk
+from tkinter import filedialog, simpledialog
+import argparse
+from ...common.utils import payload_submit, update_config, save_preset, update_size, new_random_seed
+from ...common.cn_requets import fetch_controlnet_models, progress_request, fetch_detect_image, fetch_img2img, post_request
+from ...common.output_files_utils import autosave_image, save_image
+from ...common.state import State
+
+ACCEPTED_FILE_TYPES = ["png", "jpg", "jpeg", "bmp"]
+
+# Read command-line arguments
+if __name__ == '__main__':
+    argParser = argparse.ArgumentParser()
+    argParser.add_argument("--img2img", help="img2img source file")
+
+    args = argParser.parse_args()
+
+    img2img = args.img2img
+    if img2img == '':
+        img2img = '#'  # force load file dialog if launched with --img2img without value
+
+
+state = State(img2img)
+
+if state.img2img:
+    if not os.path.exists(state.img2img):
+        root = tk.Tk()
+        root.withdraw()
+        state.img2img = filedialog.askopenfilename()
+
+    img2img_time = os.path.getmtime(state.img2img)
+
+    with Image.open(state.img2img, mode='r') as im:
+        state.render["width"] = im.width
+        state.render["height"] = im.height
+        state.render["init_width"] = state.render["width"] * 1.0
+        state.render["init_height"] = state.render["height"] * 1.0
+        update_size(state)
+
+img2img_waiting = False
+img2img_time = None
+img2img_time_prev = None
+
+rendering = False
+rendering_key = False
+instant_render = False
+image_click = False
+pause_render = False
+osd_always_on_text: str|None = None
+progress = 0.0
+
+last_detect_time = time.time()
+osd_text = None
+osd_text_display_start = None
+
+if not state.configuration["config"]['controlnet_models']:
+    fetch_controlnet_models(state)
+
+
+# Initialize Pygame
+pygame.init()
+clock = pygame.time.Clock()
+
+# Set up the display
+fullscreen = False
+screen = pygame.display.set_mode((state.render["width"] * (1 if state.img2img else 2), state.render["height"]))
+display_caption = "Sd Paint"
+pygame.display.set_caption(display_caption)
+
+# Setup text
+font = pygame.font.SysFont(None, size=24)
+font_bold = pygame.font.SysFont(None, size=24, bold=True)
+text_input = ""
+
+# Set up the drawing surface
+canvas = pygame.Surface((state.render["width"] * 2, state.render["height"]))
+pygame.draw.rect(canvas, (255, 255, 255), (0, 0, state.render["width"] * (1 if state.img2img else 2), state.render["height"]))
+
+# Set up the brush
+brush_size = {1: 2, 2: 2, 'e': 10}
+brush_colors = {
+    1: (0, 0, 0),  # Left mouse button color
+    2: (255, 255, 255),  # Middle mouse button color
+    'e': (255, 255, 255),  # Eraser color
+}
+brush_pos = {1: None, 2: None, 'e': None}
+prev_pos = None
+prev_pos2 = None
+shift_pos = None
+eraser_down = False
+render_wait = 0.5 if not state.img2img else 0.0  # wait time max between 2 draw before launching the render
+last_draw_time = time.time()
+last_render_bytes: io.BytesIO = None
+
+# Define the cursor size and color
+cursor_size = 1
+cursor_color = (0, 0, 0)
+
+
+def load_preset(state, preset_type, index):
+    """
+        Load a preset values.
+    :param str preset_type: The preset type. ``[render, controlnet]``
+    :param int index: The preset numeric keymap.
+    """
+
+    presets = state.presets["list"]
+    index = str(index)
+
+    if presets[preset_type].get(index, None) is None:
+        return f"No {preset_type} preset {index}"
+
+    preset = presets[preset_type][index]
+
+    if index == '0':
+        text = f"Load default settings:"
+
+        if preset_type == 'controlnet':
+            # prepend OSD output with render preset values for default settings display (both called successively)
+            for preset_field in state.render["preset_fields"]:
+                text += f"\n  {preset_field[:1].upper()}{preset_field[1:].replace('_', ' ')} :: {presets['render'][index][preset_field]}"
+    else:
+        text = f"Load {preset_type} preset {index}:"
+
+    # load preset
+    for preset_field in (state.render["preset_fields"] if preset_type == 'render' else state.control_net["preset_fields"]):
+        if preset.get(preset_field, None) is None:
+            continue
+        state.render[preset_field] = preset[preset_field]
+        text += f"\n  {preset_field[:1].upper()}{preset_field[1:].replace('_', ' ')} :: {preset[preset_field]}"
+
+    update_size(state)
+    return text
+
+
+def interrupt_rendering():
+    """
+        Interrupt current rendering.
+    """
+
+    global rendering, need_redraw
+
+    response = requests.post(url=f'{state.server["url"]}/sdapi/v1/interrupt')
+    if response.status_code == 200:
+        osd(text="Interrupted rendering")
+
+
+# Init the default preset
+save_preset(state, 'render', 0)
+save_preset(state, 'controlnet', 0)
+
+
+def load_filepath_into_canvas(file_path):
+    """
+        Load an image file on the sketch canvas.
+
+    :param str file_path: Local image file path.
+    """
+    global canvas
+    canvas = pygame.Surface((state.render["width"] * (1 if state.img2img else 2), state.render["height"]))
+    pygame.draw.rect(canvas, (255, 255, 255), (0, 0, state.render["width"] * (1 if state.img2img else 2), state.render["height"]))
+    img = pygame.image.load(file_path)
+    img = pygame.transform.smoothscale(img, (state.render["width"], state.render["height"]))
+    canvas.blit(img, (state.render["width"], 0))
+
+
+def finger_pos(finger_x, finger_y):
+    x = round(min(max(finger_x, 0), 1) * state.render["width"] * (1 if state.img2img else 2))
+    y = round(min(max(finger_y, 0), 1) * state.render["height"])
+    return x, y
+
+
+def save_file_dialog():
+    """
+        Display save file dialog, then write the file.
+    """
+    global last_render_bytes
+
+    root = tk.Tk()
+    root.withdraw()
+    file_path = filedialog.asksaveasfilename(defaultextension=".png")
+
+    if file_path:
+        save_image(file_path, last_render_bytes)
+        save_sketch(file_path)
+        time.sleep(1)  # add a 1-second delay
+
+
+def save_sketch(file_path):
+    file_name, file_ext = os.path.splitext(file_path)
+    sketch_img = canvas.subsurface(pygame.Rect(state.render["width"], 0, state.render["width"], state.render["height"])).copy()
+    pygame.image.save(sketch_img, f"{file_name}-sketch{file_ext}")
+
+
+def load_file_dialog():
+    """
+        Display loading file dialog, then load the image on the sketch canvas.
+    """
+
+    root = tk.Tk()
+    root.withdraw()
+    file_path = filedialog.askopenfilename()
+    if not file_path:
+        return
+
+    extension = os.path.splitext(file_path)[1][1:].lower()
+    if extension in ACCEPTED_FILE_TYPES:
+        load_filepath_into_canvas(file_path)
+
+
+def update_image(image_data):
+    """
+        Redraw the image canvas.
+
+    :param str|bytes image_data: Base64 encoded image data, from API response.
+    """
+    global last_render_bytes
+
+    # Decode base64 image data
+    if isinstance(image_data, str):
+        image_data = base64.b64decode(image_data)
+
+    img_bytes = io.BytesIO(image_data)
+    img_surface = pygame.image.load(img_bytes)
+
+    if state.autosave["images"]:
+        file_name = autosave_image(state, io.BytesIO(image_data))
+        save_sketch(file_name)
+    last_render_bytes = io.BytesIO(image_data)  # store rendered image in memory
+
+    if state.render["soft_upscale"] != 1.0:
+        img_surface = pygame.transform.smoothscale(img_surface, (img_surface.get_width() * state.render["soft_upscale"], img_surface.get_height() * state.render["soft_upscale"]))
+
+    canvas.blit(img_surface, (0, 0))
+    global need_redraw
+    need_redraw = True
+
+
+def update_batch_images(image_datas):
+    """
+        Redraw the image canvas with multiple images.
+
+    :param list[str]|list[bytes] image_datas: Images data, if ``str`` type : base64 encoded from API response.
+    """
+    global last_render_bytes
+
+    # Close old batch images
+    if len(state.render["batch_images"]):
+        for batch_image in state.render["batch_images"]:
+            image_bytes = batch_image.get('image', None)
+            if isinstance(image_bytes, io.BytesIO):
+                image_bytes.close()
+
+        state.render["batch_images"] = []
+
+    to_autosave = []
+    nb = math.ceil(math.sqrt(len(image_datas)))
+    i, j, batch_index = 0, 0, 1
+    for image_data in image_datas:
+        pos = (i * state.render["width"] // nb, j * state.render["height"] // nb)
+
+        # Decode base64 image data
+        if isinstance(image_data, str):
+            image_data = base64.b64decode(image_data)
+
+        img_bytes = io.BytesIO(image_data)
+        img_surface = pygame.image.load(img_bytes)
+
+        if (i, j) == (0, 0):
+            last_render_bytes = io.BytesIO(image_data)  # store first rendered image in memory
+
+        if state.render["soft_upscale"] != 1.0:
+            img_surface = pygame.transform.smoothscale(img_surface, (img_surface.get_width() * state.render["soft_upscale"] // nb, img_surface.get_height() * state.render["soft_upscale"] // nb))
+
+        if state.autosave["images"]:
+            to_autosave.append(io.BytesIO(image_data))
+
+        state.render["batch_images"].append({
+            "seed": state.gen_settings["seed"] + batch_index - 1,
+            "image": io.BytesIO(image_data),
+            "coord": (pos[0], pos[1], img_surface.get_width(), img_surface.get_height())
+        })
+
+        canvas.blit(img_surface, pos)
+
+        # increase indices
+        i = (i + 1) % nb
+        if i == 0:
+            j = (j + 1) % nb
+        batch_index += 1
+
+    if to_autosave:
+        file_names = autosave_image(state, to_autosave)
+        for file_name in file_names:
+            save_sketch(file_name)
+
+    global need_redraw
+    need_redraw = True
+
+
+def select_batch_image(pos):
+    """
+        Select a batch image by clicking on it.
+    :param list[int]|tuple[int] pos: The event position.
+    """
+    global need_redraw
+
+    if not len(state.render["batch_images"]):
+        return
+
+    for batch_image in state.render["batch_images"]:
+        if batch_image['coord'][0] <= pos[0] < batch_image['coord'][0] + batch_image['coord'][2] \
+                and batch_image['coord'][1] <= pos[1] < batch_image['coord'][1] + batch_image['coord'][3]:
+
+            need_redraw = True
+            osd(text_time=f"Select batch image seed {batch_image['seed']}")
+
+            if batch_image.get('image', None):
+                update_image(batch_image['image'].getbuffer().tobytes())
+
+            state.gen_settings["seed"] = batch_image['seed']
+
+            toggle_batch_mode()
+
+            break
+
+
+def img2img_submit(force=False):
+    """
+        Read the ``img2img`` file if modified since last render, check every 1s. Call the API to render if needed.
+
+    :param bool force: Force the rendering, even if the file is not modified.
+    """
+    global img2img_time, img2img_time_prev, img2img_waiting
+    img2img_waiting = False
+
+    img2img_time = os.path.getmtime(state.img2img)
+    if img2img_time != img2img_time_prev or force:
+        img2img_time_prev = img2img_time
+
+        state.server["busy"] = True
+
+        t = threading.Thread(target=progress_bar)
+        t.start()
+
+        response = fetch_img2img(state)
+        if response["status_code"] == 200:
+            return_img = response["image"]
+            update_image(return_img)
+            r_info = json.loads(response['info'])
+            return_prompt = r_info['prompt']
+            return_seed = r_info['seed']
+            global display_caption
+            display_caption = f"Sd Paint | Seed: {return_seed} | Prompt: {return_prompt}"
+        else:
+            osd(text=f"Error code returned: HTTP {response['status_code']}")
+
+        state.server["busy"] = False
+
+    if not img2img_waiting and running:
+        img2img_waiting = True
+        time.sleep(1.0)
+        img2img_submit()
+
+
+def progress_bar():
+    """
+        Update the progress bar every 0.25s
+    """
+    global progress
+
+    if not state.server["busy"]:
+        return
+
+    progress_json = progress_request(state)
+    if progress_json.get("status_code", None):
+        osd(text=f"Error code returned: HTTP {progress_json['status_code']}")
+    progress = progress_json.get('progress', None)
+    # if progress is not None and progress > 0.0:
+    #     print(f"{progress*100:.0f}%")
+
+    if state.server["busy"]:
+        time.sleep(0.25)
+        progress_bar()
+
+
+def draw_osd_text(text, rect, color=(255, 255, 255), shadow_color=(0, 0, 0), distance=1, right_align=False):
+    """
+        Draw OSD text with outline.
+    :param str text: The text to draw.
+    :param list[int]|tuple[int]|pygame.Rect rect: Destination rect.
+    :param tuple|int color: Text color.
+    :param tuple|int shadow_color: Outline color.
+    :param int distance: Outline/shadow size.
+    :param bool right_align: Align text to the right.
+    """
+
+    align_offset = 0
+    if right_align:
+        align_offset = -1
+
+    text_surface = font.render(text, True, shadow_color)
+    screen.blit(text_surface, (rect[0] + text_surface.get_width() * align_offset + distance, rect[1] + distance, rect[2], rect[3]))
+    screen.blit(text_surface, (rect[0] + text_surface.get_width() * align_offset - distance, rect[1] + distance, rect[2], rect[3]))
+    screen.blit(text_surface, (rect[0] + text_surface.get_width() * align_offset + distance, rect[1] - distance, rect[2], rect[3]))
+    screen.blit(text_surface, (rect[0] + text_surface.get_width() * align_offset - distance, rect[1] - distance, rect[2], rect[3]))
+    text_surface = font.render(text, True, color)
+    screen.blit(text_surface, (rect[0] + text_surface.get_width() * align_offset, rect[1], rect[2], rect[3]))
+
+
+def osd(**kwargs):
+    """
+        OSD display: progress bar and text messages.
+
+    :param kwargs: Accepted parameters : ``progress, text, text_time, need_redraw``
+    """
+
+    global osd_text, osd_text_display_start
+
+    osd_size = (128, 20)
+    osd_margin = 10
+    osd_progress_pos = (state.render["width"] * (0 if state.img2img else 1) + osd_margin, osd_margin)  # top left of canvas
+    # osd_pos = (width*(1 if img2img else 2) // 2 - osd_size [0] // 2, osd_margin)  # top center
+    # osd_progress_pos = (width*(1 if img2img else 2) - osd_size[0] - osd_margin, height - osd_size[1] - osd_margin)  # bottom right
+
+    osd_dot_size = osd_size[1] // 2
+    # osd_dot_pos = (width*(0 if img2img else 1) + osd_margin, osd_margin, osd_dot_size, osd_dot_size)  # top left
+    osd_dot_pos = (state.render["width"]*(1 if state.img2img else 2) - osd_dot_size * 2 - osd_margin, osd_margin, osd_dot_size, osd_dot_size)  # top right
+
+    osd_text_pos = (state.render["width"]*(0 if state.img2img else 1) + osd_margin, osd_margin)  # top left of canvas
+    # osd_text_pos = (width*(0 if img2img else 1) + osd_margin, height - osd_size[1] - osd_margin)  # bottom left of canvas
+    osd_text_offset = 0
+
+    osd_text_split_offset = 250
+
+    global progress, need_redraw, osd_always_on_text
+
+    progress = kwargs.get('progress', progress)  # type: float
+    text = kwargs.get('text', osd_text)  # type: str
+    text_time = kwargs.get('text_time', 2.0)  # type: float
+    need_redraw = kwargs.get('need_redraw', need_redraw)  # type: bool
+    osd_always_on_text = kwargs.get('always_on', osd_always_on_text)
+
+    if rendering or (state.server["busy"] and progress is not None and progress < 0.02):
+        rendering_dot_surface = pygame.Surface(osd_size, pygame.SRCALPHA)
+
+        pygame.draw.circle(rendering_dot_surface, (0, 0, 0), (osd_dot_size + 2, osd_dot_size + 2), osd_dot_size - 2)
+        pygame.draw.circle(rendering_dot_surface, (0, 200, 160), (osd_dot_size, osd_dot_size), osd_dot_size - 2)
+        screen.blit(rendering_dot_surface, osd_dot_pos)
+
+    if progress is not None and progress > 0.01:
+        need_redraw = True
+
+        # progress bar
+        progress_surface = pygame.Surface(osd_size, pygame.SRCALPHA)
+        pygame.draw.rect(progress_surface, (0, 0, 0), pygame.Rect(2, 2, math.floor(osd_size[0] * progress), osd_size[1]))
+        pygame.draw.rect(progress_surface, (0, 200, 160), pygame.Rect(0, 0, math.floor(osd_size[0] * progress), osd_size[1] - 2))
+
+        screen.blit(progress_surface, pygame.Rect(osd_progress_pos[0], osd_progress_pos[1], osd_size[0], osd_size[1]))
+
+        # progress text
+        draw_osd_text(f"{progress * 100:.0f}%", (osd_size[0] - osd_margin + osd_progress_pos[0], 3 + osd_progress_pos[1], osd_size[0], osd_size[1]), right_align=True)
+
+        osd_text_offset = osd_size[1] + osd_margin
+
+    if osd_always_on_text:
+        need_redraw = True
+
+        # OSD always-on text
+        for line in osd_always_on_text.split('\n'):
+            need_redraw = True
+
+            if '::' in line:
+                line, line_value = line.split('::')
+                line = line.rstrip(' ')
+                line_value = line_value.lstrip(' ')
+            else:
+                line_value = None
+
+            draw_osd_text(line, (osd_text_pos[0], osd_text_pos[1] + osd_text_offset, osd_size[0], osd_size[1]))
+            if line_value:
+                draw_osd_text(line_value, (osd_text_pos[0] + osd_text_split_offset, osd_text_pos[1] + osd_text_offset, osd_size[0], osd_size[1]))
+
+            osd_text_offset += osd_size[1]
+
+    if text:
+        need_redraw = True
+
+        # OSD text
+        if osd_text_display_start is None or text != osd_text:
+            osd_text_display_start = time.time()
+        osd_text = text
+
+        for line in osd_text.split('\n'):
+            need_redraw = True
+
+            if '::' in line:
+                line, line_value = line.split('::')
+                line = line.rstrip(' ')
+                line_value = line_value.lstrip(' ')
+            else:
+                line_value = None
+
+            draw_osd_text(line, (osd_text_pos[0], osd_text_pos[1] + osd_text_offset, osd_size[0], osd_size[1]))
+            if line_value:
+                draw_osd_text(line_value, (osd_text_pos[0] + osd_text_split_offset, osd_text_pos[1] + osd_text_offset, osd_size[0], osd_size[1]))
+
+            osd_text_offset += osd_size[1]
+
+        if time.time() - osd_text_display_start > text_time:
+            osd_text = None
+            osd_text_display_start = None
+
+
+def get_image_string_from_pygame():
+    img = canvas.subsurface(pygame.Rect(state.render["width"], 0, state.render["width"], state.render["height"])).copy()
+
+    if not state.render["use_invert_module"]:
+        # Convert the Pygame surface to a PIL image
+        pil_img = Image.frombytes('RGB', img.get_size(), pygame.image.tostring(img, 'RGB'))
+
+        # Invert the colors of the PIL image
+        pil_img = ImageOps.invert(pil_img)
+
+        # Convert the PIL image back to a Pygame surface
+        img = pygame.image.fromstring(pil_img.tobytes(), pil_img.size, pil_img.mode).convert_alpha()
+
+    # Save the inverted image as base64-encoded data
+    data = io.BytesIO()
+    pygame.image.save(img, data)
+    return base64.b64encode(data.getvalue()).decode('utf-8')
+
+
+def send_request():
+    """
+        Send the API request.
+
+        Use ``main_json_data`` variable.
+    """
+
+    response = post_request(state)
+    if response["status_code"] == 200:
+        if response.get("image", None):
+            update_image(response["image"])
+        elif response.get("batch_images", None):
+            update_batch_images(response["image"])
+
+        r_info = json.loads(response['info'])
+        return_prompt = r_info['prompt']
+        return_seed = r_info['seed']
+        global display_caption
+        display_caption = f"Sd Paint | Seed: {return_seed} | Prompt: {return_prompt}"
+    else:
+        osd(text=f"Error code returned: HTTP {response['status_code']}")
+    state.server["busy"] = False
+
+
+def render():
+    """
+        Call the API to launch the rendering, if another rendering is not in progress.
+    """
+    global instant_render
+
+    if time.time() - last_draw_time < render_wait and not instant_render:
+        time.sleep(0.25)
+        render()
+        return
+
+    instant_render = False
+
+    if not state.server["busy"]:
+        state.server["busy"] = True
+
+        if not state.img2img:
+            image_string = get_image_string_from_pygame()
+            payload_submit(state, image_string)
+            t = threading.Thread(target=send_request)
+            t.start()
+            t = threading.Thread(target=progress_bar)
+            t.start()
+        else:
+            t = threading.Thread(target=functools.partial(img2img_submit, True))
+            t.start()
+
+
+def get_angle(pos1, pos2):
+    """
+        Get the angle between two position.
+    :param tuple[int]|list[int] pos1: First position.
+    :param tuple[int]|list[int] pos2: Second position.
+    :param bool deg: Get the angle as degrees, otherwise radians.
+    :return: radians, degrees, cos, sin
+    """
+
+    dx = pos1[0] - pos2[0]
+    dy = pos1[1] - pos2[1]
+    rads = math.atan2(-dy, dx)
+    rads %= 2 * math.pi
+
+    return rads, math.degrees(rads), math.cos(rads), math.sin(rads)
+
+
+def brush_stroke(pos):
+    """
+        Draw the brush stroke.
+    :param tuple[int]|list[int] pos: The brush current position.
+    """
+
+    global prev_pos, prev_pos2
+
+    if prev_pos is None or (abs(event.pos[0] - prev_pos[0]) < brush_size[button] // 4 and abs(event.pos[1] - prev_pos[1]) < brush_size[button] // 4):
+        # Slow brush stroke, draw circles
+        pygame.draw.circle(canvas, brush_colors[button], event.pos, brush_size[button])
+
+    elif not prev_pos2 or brush_size[button] < 4:
+        # Draw a simple polygon for small brush sizes
+        pygame.draw.polygon(canvas, brush_colors[button], [prev_pos, event.pos], brush_size[button] * 2)
+
+    else:
+        # Draw a complex shape with gfxdraw for bigger bush sizes to avoid gaps
+        angle_prev = get_angle(prev_pos, prev_pos2)
+        angle = get_angle(event.pos, prev_pos)
+
+        offset_pos_prev = [(brush_size[button] * angle_prev[3]), (brush_size[button] * angle_prev[2])]
+        offset_pos = [(brush_size[button] * angle[3]), (brush_size[button] * angle[2])]
+        pygame.gfxdraw.filled_polygon(canvas, [
+            (prev_pos2[0] - offset_pos_prev[0], prev_pos2[1] - offset_pos_prev[1]),
+            (prev_pos[0] - offset_pos[0], prev_pos[1] - offset_pos[1]),
+            (event.pos[0] - offset_pos[0], event.pos[1] - offset_pos[1]),
+            (event.pos[0] + offset_pos[0], event.pos[1] + offset_pos[1]),
+            (prev_pos[0] + offset_pos[0], prev_pos[1] + offset_pos[1]),
+            (prev_pos2[0] + offset_pos_prev[0], prev_pos2[1] + offset_pos_prev[1])
+        ], brush_colors[button])
+
+    prev_pos2 = prev_pos
+    prev_pos = event.pos
+
+
+def shift_down():
+    return pygame.key.get_mods() & pygame.KMOD_SHIFT
+
+
+def ctrl_down():
+    return pygame.key.get_mods() & pygame.KMOD_CTRL
+
+
+def alt_down():
+    return pygame.key.get_mods() & pygame.KMOD_ALT
+
+
+
+def controlnet_detect():
+    """
+        Call ControlNet active detector on the last rendered image, replace the canvas sketch by the detector result.
+    """
+    global last_detect_time
+
+    img = canvas.subsurface(pygame.Rect(0, 0, state.render["width"], state.render["height"])).copy()
+
+    # Convert the Pygame surface to a PIL image
+    pil_img = Image.frombytes('RGB', img.get_size(), pygame.image.tostring(img, 'RGB'))
+
+    # Invert the colors of the PIL image
+    pil_img = ImageOps.invert(pil_img)
+
+    # Convert the PIL image back to a Pygame surface
+    img = pygame.image.fromstring(pil_img.tobytes(), pil_img.size, pil_img.mode).convert_alpha()
+
+    # Save the inverted image as base64-encoded data
+    data = io.BytesIO()
+    pygame.image.save(img, data)
+    data = base64.b64encode(data.getvalue()).decode('utf-8')
+
+    response = fetch_detect_image(state, data, img.get_width(), img.get_height())
+    if response["status_code"] == 200:
+        return_img = response["image"]
+        img_bytes = io.BytesIO(base64.b64decode(return_img))
+        pil_img = Image.open(img_bytes)
+        pil_img = ImageOps.invert(pil_img)
+        pil_img = pil_img.convert('RGB')
+        img_surface = pygame.image.fromstring(pil_img.tobytes(), pil_img.size, pil_img.mode)
+
+        canvas.blit(img_surface, (state.render["width"], 0))
+    else:
+        osd(text=f"Error code returned: HTTP {response['status_code']}")
+
+
+def display_configuration(wrap=True):
+    """
+        Display configuration on screen.
+    :param bool wrap: Wrap long text.
+    """
+
+    fields = [
+        '--Prompt',
+        'state/gen_settings/prompt',
+        'state/gen_settings/negative_prompt',
+        'state/gen_settings/seed',
+        '--Render',
+        'settings.steps',
+        'settings.cfg_scale',
+        'state/render/hr_scale',
+        'state/render/hr_upscaler',
+        'state/render/denoising_strength',
+        'state/render/clip_skip',
+        '--ControlNet',
+        'state/control_net/controlnet_model',
+        'state/control_net/controlnet_weight',
+        'state/control_net/controlnet_guidance_end',
+        'state/render/pixel_perfect'
+    ]
+
+    if wrap and state.render["width"] < 800:
+        wrap = 50
+    elif wrap:
+        wrap = 80
+
+    text = ''
+
+    for field in fields:
+        if field == 'settings.steps' and state.render["quick_mode"]:
+            field = 'settings.quick_steps'
+
+        # Display separator
+        if field.startswith('--'):
+            text += '\n'+field[2:]+'\n'
+            continue
+
+        # Field value
+        label = ''
+        value = ''
+
+        if '.' in field:
+            field = field.split('.')
+            var = globals().get(field[0], None)
+            if var is None:
+                continue
+
+            if isinstance(var, dict) and var.get(field[1], None) is not None:
+                label = field[1]
+                value = var.get(field[1])
+            elif (isinstance(var, list) or isinstance(var, tuple)) and field[1].isnumeric() and int(field[1]) < len(var):
+                label = field[0]
+                value = var[int(field[1])]
+            elif getattr(var, field[1], None) is not None:
+                label = field[1]
+                value = getattr(var, field[1])
+        else:
+            if 'state/' in field:
+                field_components = field.replace('state/', '').split('/')
+                field_value = getattr(state, field_components[0])[field_components[1]]
+            else:
+                field_value = globals().get(field, None)
+            if field_value is not None:
+                label = field
+                value = field_value
+
+        if label and value is not None:
+            value = str(value)
+            label = label.replace('_', ' ')
+            if label.endswith('prompt'):
+                value = value.replace(', ', ',').replace(',', ', ')  # nicer prompt display
+
+            # wrap text
+            if wrap and len(value) > wrap:
+                new_value = ''
+                to_wrap = 0
+                for i in range(len(value)):
+                    if i % wrap == wrap - 1:
+                        to_wrap = i
+
+                    if to_wrap and value[i] in [' ', ')'] or (to_wrap and i - to_wrap > 5):  # try to wrap after space
+                        new_value += value[i]+'\n::'
+                        to_wrap = 0
+                        continue
+
+                    new_value += value[i]
+
+                value = new_value
+
+            text += f"    {label} :: {value}"
+
+        text += '\n'
+
+    osd(always_on=text.strip('\n'))
+
+
+def toggle_batch_mode(cycle=False):
+    """
+        Toggle batch mode on/off. Alter the setting of HR fix if needed.
+    :param bool cycle: Cycle the batch size value.
+    """
+    global rendering_key, rendering
+
+    batch_size = state.render["batch_size"]
+    if cycle:
+        rendering_key = True
+        batch_sizes = state.render["batch_sizes"]
+        state.render["batch_size"] = batch_sizes[(batch_sizes.index(state.render["batch_size"]) + 1) % len(batch_sizes)]
+    else:
+        rendering = True
+        if state.render["batch_size"] != 1:
+            state.render["batch_size_prev"] = batch_size
+            state.render["batch_size"] = 1
+        else:
+            state.render["batch_size"] = state.render["batch_size_prev"]
+
+    if batch_size == 1:
+        state.render["hr_scale"] = state.render["batch_hr_scale_prev"]
+        update_size(state)
+        osd(text=f"Batch rendering: off")
+    else:
+        state.render["batch_hr_scale_prev"] = state.render["hr_scale"]
+        state.render["hr_scale"] = 1.0
+        update_size(state)
+        osd(text=f"Batch rendering size: {batch_size}")
+
+
+class TextDialog(simpledialog.Dialog):
+    """
+        Text input dialog.
+    """
+
+    def __init__(self, text, title, dialog_width=800, dialog_height=100):
+        self.text = text
+        self.dialog_width = dialog_width
+        self.dialog_height = dialog_height
+        super().__init__(None, title=title)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.quit()
+        gc.collect()
+
+    def body(self, master):
+        self.geometry(f'{self.dialog_width}x{self.dialog_height}')
+
+        self.e1 = tk.Text(master)
+        self.e1.insert("1.0", self.text)
+        self.e1.pack(padx=0, pady=0, fill=tk.BOTH)
+
+        self.attributes("-topmost", True)
+        master.pack(fill=tk.BOTH, expand=True)
+
+        return self.e1
+
+    def apply(self):
+        if "_"+self.e1.get("1.0", tk.INSERT)[-1:]+"_" == "_\n_":
+            p = self.e1.get("1.0", tk.INSERT)[:-1] + self.e1.get(tk.INSERT, tk.END)  # remove new line inserted when validating the dialog with ENTER
+        else:
+            p = self.e1.get("1.0", tk.END)
+        self.result = p.strip("\n")
+
+
+# Initial img2img call
+if state.img2img:
+    t = threading.Thread(target=img2img_submit)
+    t.start()
+
+# Set up the main loop
+running = True
+need_redraw = True
+
+while running:
+    rendering = False
+
+    # Handle events
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            running = False
+
+        elif event.type == pygame.MOUSEBUTTONDOWN or event.type == pygame.FINGERDOWN:
+            # Handle brush stroke start and modifiers
+            if event.type == pygame.FINGERDOWN:
+                event.button = 1
+                event.pos = finger_pos(event.x, event.y)
+
+            if event.pos[0] < state.render["width"]:
+                image_click = True  # clicked on the image part
+                if state.render["batch_size"] != 1 and len(state.render["batch_images"]):
+                    select_batch_image(event.pos)
+            else:
+                need_redraw = True
+                last_draw_time = time.time()
+
+                brush_key = event.button
+                if eraser_down:
+                    brush_key = 'e'
+
+                if brush_key in brush_colors:
+                    brush_pos[brush_key] = event.pos
+                elif event.button == 4:  # scroll up
+                    brush_size[1] = max(1, brush_size[1] + 1)
+                    brush_size[2] = max(1, brush_size[2] + 1)
+                    osd(text=f"Brush size {brush_size[1]}")
+
+                elif event.button == 5:  # scroll down
+                    brush_size[1] = max(1, brush_size[1] - 1)
+                    brush_size[2] = max(1, brush_size[2] - 1)
+                    osd(text=f"Brush size {brush_size[1]}")
+
+                if shift_down() and brush_pos[brush_key] is not None:
+                    if shift_pos is None:
+                        shift_pos = brush_pos[brush_key]
+                    else:
+                        pygame.draw.polygon(canvas, brush_colors[brush_key], [shift_pos, brush_pos[brush_key]], brush_size[brush_key] * 2)
+                        shift_pos = brush_pos[brush_key]
+
+        elif event.type == pygame.MOUSEBUTTONUP or event.type == pygame.FINGERUP:
+            # Handle brush stoke end
+            last_draw_time = time.time()
+
+            if event.type == pygame.FINGERUP:
+                event.button = 1
+                event.pos = finger_pos(event.x, event.y)
+
+            if not image_click:
+                need_redraw = True
+                rendering = True
+
+                if event.button in brush_colors or eraser_down:
+                    brush_key = event.button
+                    if eraser_down:
+                        brush_key = 'e'
+
+                    if brush_size[brush_key] >= 4 and getattr(event, 'pos', None) is not None:
+                        pygame.draw.circle(canvas, brush_colors[brush_key], event.pos, brush_size[brush_key])
+
+                    brush_pos[brush_key] = None
+                    prev_pos = None
+                    brush_color = brush_colors[brush_key]
+
+            image_click = False  # reset image click detection
+
+        elif event.type == pygame.MOUSEMOTION or event.type == pygame.FINGERMOTION:
+            # Handle drawing brush strokes
+            if event.type == pygame.FINGERMOTION:
+                event.pos = finger_pos(event.x, event.y)
+
+            if not image_click:
+                need_redraw = True
+                for button, pos in brush_pos.items():
+                    if pos is not None and button in brush_colors:
+                        last_draw_time = time.time()
+                        brush_stroke(pos)  # do the brush stroke
+
+        elif event.type == pygame.KEYDOWN:
+            # DBG key & modifiers
+            # print(f"key down {event.key}, modifiers:")
+            # print(f" shift:      {pygame.key.get_mods() & pygame.KMOD_SHIFT}")
+            # print(f" ctrl:       {pygame.key.get_mods() & pygame.KMOD_CTRL}")
+            # print(f" alt:        {pygame.key.get_mods() & pygame.KMOD_ALT}")
+
+            # Handle keyboard shortcuts
+            need_redraw = True
+            last_draw_time = time.time()
+            rendering = False
+
+            event.button = 1
+            if event.key == pygame.K_UP:
+                rendering = True
+                instant_render = True
+                state.gen_settings["seed"] = state.gen_settings["seed"] + state.render["batch_size"]
+                update_config(state.json_file, write=state.autosave["seed"], values={'seed': state.gen_settings["seed"]})
+                osd(text=f"Seed: {state['gen_settings']['seed']}")
+
+            elif event.key == pygame.K_DOWN:
+                rendering = True
+                instant_render = True
+                state.gen_settings["seed"] = state.gen_settings["seed"] - state.render["batch_size"]
+                update_config(state.json_file, write=state.autosave["seed"], values={'seed': state.gen_settings["seed"]})
+                osd(text=f"Seed: {state['gen_settings']['seed']}")
+
+            elif event.key == pygame.K_n:
+                if ctrl_down():
+                    with TextDialog(state.gen_settings["seed"], title="Seed", dialog_width=200, dialog_height=30) as dialog:
+                        if dialog.result and dialog.result.isnumeric():
+                            osd(text=f"Seed: {dialog.result}")
+                            state.gen_settings["seed"] = int(dialog.result)
+                            update_config(state.json_file, write=state.autosave["seed"], values={'seed': state.gen_settings["seed"]})
+                            rendering = True
+                            instant_render = True
+                else:
+                    rendering = True
+                    instant_render = True
+                    state.gen_settings["seed"] = new_random_seed()
+                    update_config(state.json_file, write=state.autosave["seed"], values={'seed': state.gen_settings["seed"]})
+                    osd(text=f"Seed: {state['gen_settings']['seed']}")
+
+            elif event.key == pygame.K_c:
+                if shift_down():
+                    rendering_key = True
+                    state.render["clip_skip"] -= 1
+                    state.render["clip_skip"] = (state.render["clip_skip"] + 1) % 2
+                    state.render["clip_skip"] += 1
+                    osd(text=f"CLIP skip: {state['render']['clip_skip']}")
+                else:
+                    display_configuration()
+
+            elif event.key == pygame.K_m and state.control_net["controlnet_models"]:
+                if shift_down():
+                    rendering_key = True
+                    controlnet_models = state.control_net["controlnet_models"]
+                    state.control_net["controlnet_models"] = controlnet_models[(controlnet_models.index(state.control_net["controlnet_models"]) + 1) % len(controlnet_models)]
+                    osd(text=f"ControlNet model: {state['control_net']['controlnet_models']}")
+
+            elif event.key == pygame.K_i:
+                if ctrl_down():
+                    interrupt_rendering()
+            elif event.key == pygame.K_h:
+                if shift_down():
+                    rendering_key = True
+                    state.render["hr_scale"] = state.render["hr_scales"][(state.render["hr_scales"].index(state.render["hr_scale"])+1) % len(state.render["hr_scales"])]
+                else:
+                    rendering = True
+                    if state.render["hr_scale"] != 1.0:
+                        state.render["hr_scale_prev"] = state.render["hr_scale"]
+                        state.render["hr_scale"] = 1.0
+                    else:
+                        state.render["hr_scale"] = state.render["hr_scale_prev"]
+
+                if state.render["hr_scale"] == 1.0:
+                    osd(text="HR scale: off")
+                else:
+                    osd(text=f"HR scale: {state['hr']['hr_scale']}")
+
+                update_size(state, hr_scale=state.render["hr_scale"])
+
+            elif event.key in (pygame.K_KP_ENTER, pygame.K_RETURN):
+                rendering = True
+                instant_render = True
+                osd(text=f"Rendering")
+
+            elif event.key == pygame.K_q:
+                rendering = True
+                instant_render = True
+                state.render["quick_mode"] = not state.render["quick_mode"]
+                if state.render["quick_mode"]:
+                    osd(text=f"Quick render: on")
+                    state.render["hr_scale_prev"] = state.render["hr_scale"]
+                    state.render["hr_scale"] = 1.0
+                else:
+                    osd(text=f"Quick render: off")
+                    state.render["hr_scale"] = state.render["hr_scale_prev"]
+
+                update_size(state, hr_scale=state.render["hr_scale"])
+
+            elif event.key == pygame.K_a:
+                state.autosave["images"] = not state.autosave["images"]
+                osd(text=f"Autosave images: {'on' if state['autosave']['images'] else 'off'}")
+
+            elif event.key == pygame.K_o:
+                if ctrl_down():
+                    rendering = True
+                    instant_render = True
+                    load_file_dialog()
+
+            elif event.key in (pygame.K_KP0, pygame.K_KP1, pygame.K_KP2, pygame.K_KP3, pygame.K_KP4,
+                               pygame.K_KP5, pygame.K_KP6, pygame.K_KP7, pygame.K_KP8, pygame.K_KP9):
+                keymap = {
+                    pygame.K_KP0:   0,
+                    pygame.K_KP1:   1,
+                    pygame.K_KP2:   2,
+                    pygame.K_KP3:   3,
+                    pygame.K_KP4:   4,
+                    pygame.K_KP5:   5,
+                    pygame.K_KP6:   6,
+                    pygame.K_KP7:   7,
+                    pygame.K_KP8:   8,
+                    pygame.K_KP9:   9
+                }
+
+                if ctrl_down():
+                    if event.key != pygame.K_KP0:
+                        preset_info = save_preset(state, 'controlnet' if alt_down() else 'render', keymap.get(event.key))
+                        if preset_info['index'] == '0':
+                            osd(text=f"Save {preset_info['preset_type']} preset {preset_info['index']}")
+                else:
+                    rendering = True
+                    instant_render = True
+
+                    if event.key == pygame.K_KP0:
+                        # Reset both render & controlnet settings if keypad 0
+                        text = load_preset(state, 'render', keymap.get(event.key))
+                        osd(text=text, text_time=4.0)
+                        text = load_preset(state, 'controlnet', keymap.get(event.key))
+                        osd(text=text, text_time=4.0)
+                    else:
+                        text = load_preset(state, 'controlnet' if alt_down() else 'render', keymap.get(event.key))
+                        osd(text=text, text_time=4.0)
+
+            elif event.key == pygame.K_p:
+                if ctrl_down():
+                    pause_render = not pause_render
+
+                    if pause_render:
+                        osd(text=f"On-demand rendering (ENTER to render)")
+
+                    else:
+                        rendering = True
+                        instant_render = True
+                        osd(text=f"Dynamic rendering")
+                elif alt_down():
+                    with TextDialog(state.gen_settings["negative_prompt"], title="Negative prompt") as dialog:
+                        if dialog.result:
+                            osd(text=f"New negative prompt: {dialog.result}")
+                            state.gen_settings["negative_prompt"] = dialog.result
+                            update_config(state.json_file, write=state.autosave["negative_prompt"], values={'negative_prompt': state.gen_settings["negative_prompt"]})
+                            rendering = True
+                            instant_render = True
+                else:
+                    with TextDialog(state.gen_settings["prompt"], title="Prompt") as dialog:
+                        if dialog.result:
+                            osd(text=f"New prompt: {dialog.result}")
+                            state.gen_settings["prompt"] = dialog.result
+                            update_config(state.json_file, write=state.autosave["prompt"], values={'prompt': state.gen_settings["prompt"]})
+                            rendering = True
+                            instant_render = True
+
+            elif event.key == pygame.K_BACKSPACE:
+                pygame.draw.rect(canvas, (255, 255, 255), (state.render["width"], 0, state.render["width"], state.render["height"]))
+
+            elif event.key == pygame.K_s:
+                if ctrl_down():
+                    save_file_dialog()
+                elif shift_down():
+                    rendering_key = True
+                    samplers = state.samplers["list"]
+                    state.samplers["sampler"] = samplers[(samplers.index(state.samplers["sampler"]) + 1) % len(samplers)]
+                    osd(text=f"Sampler: {state['samplers']['sampler']}")
+
+            elif event.key == pygame.K_e:
+                eraser_down = True
+
+            elif event.key == pygame.K_t:
+                if shift_down():
+                    if render_wait == 2.0:
+                        render_wait = 0.0
+                        osd(text="Render wait: off")
+                    else:
+                        render_wait += 0.5
+                        osd(text=f"Render wait: {render_wait}s")
+
+            elif event.key == pygame.K_u:
+                if shift_down():
+                    rendering_key = True
+                    hr_upscalers = state.render["hr_upscalers"]
+                    hr_upscaler = state.render["hr_upscaler"]
+                    state.render["hr_upscaler"] = hr_upscalers[(hr_upscalers.index(hr_upscaler) + 1) % len(hr_upscalers)]
+                    osd(text=f"HR upscaler: {hr_upscaler}")
+
+            elif event.key == pygame.K_b:
+                toggle_batch_mode(cycle=shift_down())
+
+            elif event.key == pygame.K_w:
+                if shift_down():
+                    rendering_key = True
+                    controlnet_weights = state.control_net["controlnet_weights"]
+                    controlnet_weight = state.control_net["controlnet_weight"]
+                    state.control_net["controlnet_weight"] = controlnet_weights[(controlnet_weights.index(controlnet_weight) + 1) % len(controlnet_weights)]
+                    osd(text=f"ControlNet weight: {controlnet_weight}")
+
+            elif event.key == pygame.K_g:
+                if shift_down() and ctrl_down():
+                    rendering_key = True
+                    state.render["pixel_perfect"] = not state.render["pixel_perfect"]
+                    osd(text=f"ControlNet pixel perfect mode: {'on' if state.render['pixel_perfect'] else 'off'}")
+                elif shift_down():
+                    rendering_key = True
+                    controlnet_guidance_ends = state.control_net["controlnet_guidance_ends"]
+                    controlnet_guidance_end = state.control_net["controlnet_guidance_end"]
+                    state.control_net["controlnet_guidance_end"] = controlnet_guidance_ends[(controlnet_guidance_ends.index(controlnet_guidance_end) + 1) % len(controlnet_guidance_ends)]
+                    osd(text=f"ControlNet guidance end: {controlnet_guidance_end}")
+
+            elif event.key == pygame.K_d:
+                if shift_down():
+                    rendering_key = True
+                    denoising_strength = state.render["denoising_strength"]
+                    denoising_strengths = state.render["denoising_strengths"]
+                    state.render["denoising_strength"] = denoising_strengths[(denoising_strengths.index(denoising_strength) + 1) % len(denoising_strengths)]
+                    if state.img2img:
+                        osd(text=f"Denoising: {denoising_strength}")
+                    else:
+                        osd(text=f"HR denoising: {denoising_strength}")
+                elif ctrl_down():
+                    osd(text=f"Detect {state.detectors['detector']}")
+
+                    t = threading.Thread(target=controlnet_detect())
+                    t.start()
+
+                    # select next detector
+                    state.detectors["detector"] = state.detectors["list"][(state.detectors["list"].index(state.detectors["detector"])+1) % len(state.detectors["list"])]
+
+            elif event.key == pygame.K_f:
+                fullscreen = not fullscreen
+                if fullscreen:
+                    screen = pygame.display.set_mode((0, 0), pygame.FULLSCREEN)
+                else:
+                    screen = pygame.display.set_mode((state.render["width"]*2, state.render["height"]))
+
+            elif event.key in (pygame.K_ESCAPE, pygame.K_x):
+                running = False
+                pygame.quit()
+                exit(0)
+
+        elif event.type == pygame.KEYUP:
+            # Handle special keys release
+            if event.key == pygame.K_e:
+                eraser_down = False
+                brush_pos['e'] = None
+
+            elif event.key in (pygame.K_LSHIFT, pygame.K_RSHIFT):
+                if rendering_key:
+                    rendering = True
+                    rendering_key = False
+                shift_pos = None
+
+            elif event.key in (pygame.K_c,):
+                # Remove OSD always-on text
+                need_redraw = True
+                osd(always_on=None)
+
+    # Call image render
+    if (rendering and not pause_render) or instant_render:
+        t = threading.Thread(target=render)
+        t.start()
+
+    # Draw the canvas and brushes on the screen
+    screen.blit(canvas, (0, 0))
+
+    # Create a new surface with a circle
+    cursor_size = brush_size[1]*2
+    cursor_surface = pygame.Surface((cursor_size, cursor_size), pygame.SRCALPHA)
+    pygame.draw.circle(cursor_surface, cursor_color, (cursor_size // 2, cursor_size // 2), cursor_size // 2)
+
+    # Blit the cursor surface onto the screen surface at the position of the mouse
+    mouse_pos = pygame.mouse.get_pos()
+    screen.blit(cursor_surface, (mouse_pos[0] - cursor_size // 2, mouse_pos[1] - cursor_size // 2))
+
+    # Handle OSD
+    osd()
+
+    # Update the display
+    if need_redraw:
+        pygame.display.flip()
+        pygame.display.set_caption(display_caption)
+        need_redraw = False
+
+    # Set max FPS
+    clock.tick(120)
+
+# Clean up Pygame
+pygame.quit()
